requirements.txt
requests
urllib3
coinbase-advanced-py
python-dotenv
pytz
num2words